Refactor to make the code easy to read and maintain
