+++ src/index.html
@@ -13,7 +13,7 @@
     crossorigin="anonymous" referrerpolicy="no-referrer" />
 </head>
 
-<body>
+<body id="body">
   <header>
     <div class="logo">Tv Shows logo</div>
     <ul>
